Splitt ut journalpost og endre filtype
